+++ config/mapconverter.go
@@ -24,11 +24,3 @@ type MapConverter interface {
 	// Convert applies the conversion logic to the given "cfgMap".
 	Convert(ctx context.Context, cfgMap *Map) error
 }
-
-// Deprecated: Implement MapConverter interface.
-type MapConverterFunc func(context.Context, *Map) error
-
-// Convert implements MapConverter.Convert func.
-func (f MapConverterFunc) Convert(ctx context.Context, cfgMap *Map) error {
-	return f(ctx, cfgMap)
-}